+++ scripts/generate_cdef.py
@@ -10,16 +10,16 @@ DEFAULT_HEADER = ROOT.parent / "aeron" / "aeron-client" / "src" / "main" / "c" /
 DEFAULT_OUTPUT = ROOT / "pyaeron" / "_generated_cdef.py"
 
 PRIMITIVES = """typedef _Bool bool;
-typedef long int ssize_t;
+typedef signed long long ssize_t;
 typedef unsigned char uint8_t;
 typedef signed char int8_t;
 typedef signed short int16_t;
 typedef unsigned short uint16_t;
 typedef signed int int32_t;
 typedef unsigned int uint32_t;
-typedef signed long int int64_t;
-typedef unsigned long int uint64_t;
-typedef unsigned long size_t;
+typedef signed long long int64_t;
+typedef unsigned long long uint64_t;
+typedef unsigned long long size_t;
 """
 
 REQUIRED_DECLARATIONS = {
